Create storage client service
The storage healthcheck also needs to make use of the correct storage
client and config along with the publishing of metadata.

Create a storage client service in order to use it in both places.
