+++ app/models/healthcheck/storage_check.rb
@@ -1,7 +1,5 @@
 module Healthcheck
   class StorageCheck
-    include Utilities::Configuration::Settings
-
     def name
       :storage_connectivity
     end
@@ -9,18 +7,11 @@ module Healthcheck
     def status
       healthcheck_key = 'healthcheck.txt'
 
-      unless service.exist?(healthcheck_key)
-        service.upload(healthcheck_key, '')
+      unless SelfService.service(:storage_client).exist?(healthcheck_key)
+        SelfService.service(:storage_client).upload(healthcheck_key, '')
       end
 
       OK
-    end
-
-    def service
-      @service ||= ActiveStorage::Service.configure(
-        Rails.configuration.active_storage.service,
-        configuration('storage.yml')
-      )
     end
   end
 end
